requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
Pillow==8.1.1
requests==2.25.1
torch==1.8.0+cpu
torchvision==0.9.0+cpu
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
